LayoutTests/fast/forms/onchange-change-type.html

<html>
<head>
    <script src="../../resources/js-test-pre.js"></script>
    <script>
    function test() {
        var tf = document.getElementById('tf');

        var didFireOnChange = false;
        tf.onchange = function() {
          didFireOnChange = true;
        }

        if (window.testRunner) {
            testRunner.dumpAsText();
        }

        tf.focus();
        document.execCommand('InsertText', false, 'input value');
        tf.blur();

        debug('Should fire change event when type does not change.');
        shouldBeTrue(didFireOnChange.toString());

        didFireOnChange = false;
        tf.focus();
        document.execCommand('InsertText', false, 'new input value');
        tf.setAttribute('type', 'password');
        tf.blur();

        debug('Should *not* fire change event when type does not change.');
        shouldBeFalse(didFireOnChange.toString());
    }
    </script>
</head>
<body onload="test()">
    This tests that onchange is not fired after a form field's type changes.<br>
    <form name="fm">
        <input type="text" id="tf" />
    </form>
    <div id="console"></div>
</body>
</html>
