third_party/blink/web_tests/fast/forms/onchange-change-type-expected.txt

This tests that onchange is not fired after a form field's type changes.

PASS successfullyParsed is true

TEST COMPLETE

Should fire change event when type does not change.
PASS true is true
Should *not* fire change event when type does not change.
PASS false is false
